# Use go slices