--- a/reader.go
+++ b/reader.go
@@ -7,33 +7,41 @@ package gozstd
 #include "zstd.h"
 #include "zstd_errors.h"
 
-#include <stdlib.h>  // for malloc/free
-#include <stdint.h>  // for uintptr_t
+typedef struct {
+	size_t dstSize;
+	size_t srcSize;
+	size_t dstPos;
+	size_t srcPos;
+} ZSTD_EXT_BufferSizes;
 
 // The following *_wrapper functions allow avoiding memory allocations
 // durting calls from Go.
 // See https://github.com/golang/go/issues/24450 .
 
-static size_t ZSTD_initDStream_usingDDict_wrapper(uintptr_t ds, uintptr_t dict) {
+static size_t ZSTD_initDStream_usingDDict_wrapper(void *ds, void *dict) {
     return ZSTD_initDStream_usingDDict((ZSTD_DStream*)ds, (ZSTD_DDict*)dict);
 }
 
-static size_t ZSTD_freeDStream_wrapper(uintptr_t ds) {
+static size_t ZSTD_freeDStream_wrapper(void *ds) {
     return ZSTD_freeDStream((ZSTD_DStream*)ds);
 }
 
-static size_t ZSTD_decompressStream_wrapper(uintptr_t ds, uintptr_t output, uintptr_t input) {
-    return ZSTD_decompressStream((ZSTD_DStream*)ds, (ZSTD_outBuffer*)output, (ZSTD_inBuffer*)input);
+static size_t ZSTD_decompressStream_wrapper(void *ds, void* dst, const void* src, ZSTD_EXT_BufferSizes* sizes) {
+    return ZSTD_decompressStream_simpleArgs((ZSTD_DStream*)ds, dst, sizes->dstSize, &sizes->dstPos, src, sizes->srcSize, &sizes->srcPos);
 }
 */
 import "C"
 
 import (
+	"bytes"
 	"fmt"
 	"io"
+	"reflect"
 	"runtime"
 	"unsafe"
 )
+
+const minDirectWriteBufferSize = 16 * 1024
 
 var (
 	dstreamInBufSize  = C.ZSTD_DStreamInSize()
@@ -46,11 +54,17 @@ type Reader struct {
 	ds *C.ZSTD_DStream
 	dd *DDict
 
-	inBuf  *C.ZSTD_inBuffer
-	outBuf *C.ZSTD_outBuffer
+	inBufWrapper  *bytes.Buffer
+	outBufWrapper *bytes.Buffer
 
-	inBufGo  cMemPtr
-	outBufGo cMemPtr
+	skipNextRead bool
+
+	readerPos int
+	inBuf     []byte
+	outBuf    []byte
+	// go doesn't allow passing pointers to structs with pointers to Go memory
+	// so we can't use ZSTD_inBuffer and ZSTD_outBuffer directly
+	sizes C.ZSTD_EXT_BufferSizes
 }
 
 // NewReader returns new zstd reader reading compressed data from r.
@@ -68,26 +82,18 @@ func NewReaderDict(r io.Reader, dd *DDict) *Reader {
 	ds := C.ZSTD_createDStream()
 	initDStream(ds, dd)
 
-	inBuf := (*C.ZSTD_inBuffer)(C.malloc(C.sizeof_ZSTD_inBuffer))
-	inBuf.src = C.malloc(dstreamInBufSize)
-	inBuf.size = 0
-	inBuf.pos = 0
-
-	outBuf := (*C.ZSTD_outBuffer)(C.malloc(C.sizeof_ZSTD_outBuffer))
-	outBuf.dst = C.malloc(dstreamOutBufSize)
-	outBuf.size = 0
-	outBuf.pos = 0
+	inBufWrapper := decInBufPool.Get().(*bytes.Buffer)
+	outBufWrapper := decOutBufPool.Get().(*bytes.Buffer)
 
 	zr := &Reader{
-		r:      r,
-		ds:     ds,
-		dd:     dd,
-		inBuf:  inBuf,
-		outBuf: outBuf,
+		r:             r,
+		ds:            ds,
+		dd:            dd,
+		inBufWrapper:  inBufWrapper,
+		outBufWrapper: outBufWrapper,
+		inBuf:         inBufWrapper.Bytes(),
+		outBuf:        outBufWrapper.Bytes(),
 	}
-
-	zr.inBufGo = cMemPtr(zr.inBuf.src)
-	zr.outBufGo = cMemPtr(zr.outBuf.dst)
 
 	runtime.SetFinalizer(zr, freeDStream)
 	return zr
@@ -95,10 +101,10 @@ func NewReaderDict(r io.Reader, dd *DDict) *Reader {
 
 // Reset resets zr to read from r using the given dictionary dd.
 func (zr *Reader) Reset(r io.Reader, dd *DDict) {
-	zr.inBuf.size = 0
-	zr.inBuf.pos = 0
-	zr.outBuf.size = 0
-	zr.outBuf.pos = 0
+	zr.readerPos = 0
+	zr.sizes = C.ZSTD_EXT_BufferSizes{}
+	zr.inBuf = zr.inBuf[:0]
+	zr.outBuf = zr.outBuf[:0]
 
 	zr.dd = dd
 	initDStream(zr.ds, zr.dd)
@@ -111,9 +117,7 @@ func initDStream(ds *C.ZSTD_DStream, dd *DDict) {
 	if dd != nil {
 		ddict = dd.p
 	}
-	result := C.ZSTD_initDStream_usingDDict_wrapper(
-		C.uintptr_t(uintptr(unsafe.Pointer(ds))),
-		C.uintptr_t(uintptr(unsafe.Pointer(ddict))))
+	result := C.ZSTD_initDStream_usingDDict_wrapper(unsafe.Pointer(ds), unsafe.Pointer(ddict))
 	ensureNoError("ZSTD_initDStream_usingDDict", result)
 }
 
@@ -129,21 +133,23 @@ func (zr *Reader) Release() {
 		return
 	}
 
-	result := C.ZSTD_freeDStream_wrapper(
-		C.uintptr_t(uintptr(unsafe.Pointer(zr.ds))))
+	result := C.ZSTD_freeDStream_wrapper(unsafe.Pointer(zr.ds))
 	ensureNoError("ZSTD_freeDStream", result)
 	zr.ds = nil
 
-	C.free(zr.inBuf.src)
-	C.free(unsafe.Pointer(zr.inBuf))
-	zr.inBuf = nil
-
-	C.free(zr.outBuf.dst)
-	C.free(unsafe.Pointer(zr.outBuf))
-	zr.outBuf = nil
-
 	zr.r = nil
 	zr.dd = nil
+
+	if zr.inBuf != nil {
+		zr.inBuf = nil
+		decInBufPool.Put(zr.inBufWrapper)
+		zr.inBufWrapper = nil
+	}
+	if zr.outBuf != nil {
+		zr.outBuf = nil
+		decOutBufPool.Put(zr.outBufWrapper)
+		zr.outBufWrapper = nil
+	}
 }
 
 // WriteTo writes all the data from zr to w.
@@ -152,16 +158,17 @@ func (zr *Reader) Release() {
 func (zr *Reader) WriteTo(w io.Writer) (int64, error) {
 	nn := int64(0)
 	for {
-		if zr.outBuf.pos == zr.outBuf.size {
-			if err := zr.fillOutBuf(); err != nil {
+		if zr.readerPos >= len(zr.outBuf) {
+			if _, err := zr.fillOutBuf(nil); err != nil {
 				if err == io.EOF {
 					return nn, nil
 				}
 				return nn, err
 			}
+			zr.readerPos = 0
 		}
-		n, err := w.Write(zr.outBufGo[zr.outBuf.pos:zr.outBuf.size])
-		zr.outBuf.pos += C.size_t(n)
+		n, err := w.Write(zr.outBuf[zr.readerPos:])
+		zr.readerPos += n
 		nn += int64(n)
 		if err != nil {
 			return nn, err
@@ -175,51 +182,68 @@ func (zr *Reader) Read(p []byte) (int, error) {
 		return 0, nil
 	}
 
-	if zr.outBuf.pos == zr.outBuf.size {
-		if err := zr.fillOutBuf(); err != nil {
+	if zr.readerPos >= len(zr.outBuf) {
+		if len(p) >= minDirectWriteBufferSize {
+			// write directly into the target buffer
+			// but make sure to override its capacity
+			return zr.fillOutBuf(p[:len(p):len(p)])
+		}
+		if _, err := zr.fillOutBuf(nil); err != nil {
+			return 0, err
+		}
+		zr.readerPos = 0
+	}
+
+	n := copy(p, zr.outBuf[zr.readerPos:])
+	zr.readerPos += n
+	return n, nil
+}
+
+func (zr *Reader) fillOutBuf(target []byte) (int, error) {
+	dst := target
+	if dst == nil {
+		dst = zr.outBuf
+	}
+
+	if int(zr.sizes.srcPos) == len(zr.inBuf) && !zr.skipNextRead {
+		// inBuf is empty and the previously decompressed data size
+		// is smaller than the maximum possible dst.size.
+		// This means that the internal buffer in zr.ds doesn't contain
+		// more data to decompress, so read new data into inBuf.
+		if err := zr.fillInBuf(); err != nil {
 			return 0, err
 		}
 	}
 
-	n := copy(p, zr.outBufGo[zr.outBuf.pos:zr.outBuf.size])
-	zr.outBuf.pos += C.size_t(n)
-	return n, nil
-}
+	zr.sizes.dstSize = C.size_t(cap(dst))
+	zr.sizes.dstPos = 0
 
-func (zr *Reader) fillOutBuf() error {
-	if zr.inBuf.pos == zr.inBuf.size && zr.outBuf.size < dstreamOutBufSize {
-		// inBuf is empty and the previously decompressed data size
-		// is smaller than the maximum possible zr.outBuf.size.
-		// This means that the internal buffer in zr.ds doesn't contain
-		// more data to decompress, so read new data into inBuf.
-		if err := zr.fillInBuf(); err != nil {
-			return err
-		}
-	}
-
+	inHdr := (*reflect.SliceHeader)(unsafe.Pointer(&zr.inBuf))
+	outHdr := (*reflect.SliceHeader)(unsafe.Pointer(&dst))
 tryDecompressAgain:
-	// Try decompressing inBuf into outBuf.
-	zr.outBuf.size = dstreamOutBufSize
-	zr.outBuf.pos = 0
-	prevInBufPos := zr.inBuf.pos
-	result := C.ZSTD_decompressStream_wrapper(
-		C.uintptr_t(uintptr(unsafe.Pointer(zr.ds))),
-		C.uintptr_t(uintptr(unsafe.Pointer(zr.outBuf))),
-		C.uintptr_t(uintptr(unsafe.Pointer(zr.inBuf))))
-	zr.outBuf.size = zr.outBuf.pos
-	zr.outBuf.pos = 0
+	zr.sizes.srcSize = C.size_t(len(zr.inBuf))
+	prevInBufPos := zr.sizes.srcPos
 
-	if C.ZSTD_getErrorCode(result) != 0 {
-		return fmt.Errorf("cannot decompress data: %s", errStr(result))
+	// Try decompressing inBuf into outBuf.
+	result := C.ZSTD_decompressStream_wrapper(
+		unsafe.Pointer(zr.ds), unsafe.Pointer(outHdr.Data), unsafe.Pointer(inHdr.Data), &zr.sizes)
+
+	zr.skipNextRead = int(zr.sizes.dstPos) == cap(dst)
+	if target == nil {
+		zr.outBuf = zr.outBuf[:zr.sizes.dstPos]
 	}
 
-	if zr.outBuf.size > 0 {
+	if zstdIsError(result) {
+		return int(zr.sizes.dstPos), fmt.Errorf("cannot decompress data: %s", errStr(result))
+	}
+
+	if zr.sizes.dstPos > 0 {
 		// Something has been decompressed to outBuf. Return it.
-		return nil
+		return int(zr.sizes.dstPos), nil
 	}
 
 	// Nothing has been decompressed from inBuf.
-	if zr.inBuf.pos != prevInBufPos && zr.inBuf.pos < zr.inBuf.size {
+	if zr.sizes.srcPos != prevInBufPos && int(zr.sizes.srcPos) < len(zr.inBuf) {
 		// Data has been consumed from inBuf, but decompressed
 		// into nothing. There is more data in inBuf, so try
 		// decompressing it again.
@@ -230,21 +254,25 @@ tryDecompressAgain:
 	// decompressed into nothing and inBuf became empty.
 	// Read more data into inBuf and try decompressing again.
 	if err := zr.fillInBuf(); err != nil {
-		return err
+		return 0, err
 	}
+
 	goto tryDecompressAgain
 }
 
 func (zr *Reader) fillInBuf() error {
 	// Copy the remaining data to the start of inBuf.
-	copy(zr.inBufGo[:dstreamInBufSize], zr.inBufGo[zr.inBuf.pos:zr.inBuf.size])
-	zr.inBuf.size -= zr.inBuf.pos
-	zr.inBuf.pos = 0
+	if zr.sizes.srcPos > 0 && int(zr.sizes.srcPos) > cap(zr.inBuf)/2 {
+		copy(zr.inBuf[:cap(zr.inBuf)], zr.inBuf[zr.sizes.srcPos:])
+		zr.inBuf = zr.inBuf[:len(zr.inBuf)-int(zr.sizes.srcPos)]
+		zr.sizes.srcPos = 0
+	}
 
 readAgain:
 	// Read more data into inBuf.
-	n, err := zr.r.Read(zr.inBufGo[zr.inBuf.size:dstreamInBufSize])
-	zr.inBuf.size += C.size_t(n)
+	n, err := zr.r.Read(zr.inBuf[len(zr.inBuf):cap(zr.inBuf)])
+	zr.inBuf = zr.inBuf[:len(zr.inBuf)+n]
+
 	if err == nil {
 		if n == 0 {
 			// Nothing has been read. Try reading data again.
@@ -260,5 +288,5 @@ readAgain:
 		// Do not wrap io.EOF, so the caller may notify the end of stream.
 		return err
 	}
-	return fmt.Errorf("cannot read data from the underlying reader: %s", err)
+	return fmt.Errorf("cannot read data from the underlying reader: %w", err)
 }