… writer.go