I've rewritten how the CGO wrapper is done to achieve two things:
- stop pretending that the pointers passed to C aren't real pointers, so that go can properly adjust them during stack moves (ie stop using the uintptr_t hack, which allows use of stack-allocated buffers for Compress() calls)
- pass go slices directly to the C calls, which means the buffers don't need to be copied, making things a bit faster

Here's benchmark results against master (run on M1 using go 1.18):

<details>
  <summary>CPU time</summary>

```
name                                          old time/op    new time/op    delta
DecompressDict/blockSize_1/level_3-8            21.3ns ± 2%    25.5ns ± 2%    +19.57%  (p=0.016 n=4+5)
DecompressDict/blockSize_1/level_5-8            21.0ns ± 2%    25.2ns ± 2%    +20.02%  (p=0.016 n=4+5)
DecompressDict/blockSize_1/level_10-8           21.3ns ± 5%    25.4ns ± 3%    +19.18%  (p=0.016 n=4+5)
DecompressDict/blockSize_10/level_3-8           20.8ns ± 4%    25.2ns ± 1%    +20.92%  (p=0.016 n=4+5)
DecompressDict/blockSize_10/level_5-8           20.9ns ± 3%    25.8ns ± 2%    +23.59%  (p=0.016 n=4+5)
DecompressDict/blockSize_10/level_10-8          26.7ns ± 2%    31.2ns ± 1%    +16.90%  (p=0.016 n=4+5)
DecompressDict/blockSize_100/level_3-8          26.7ns ± 0%    32.1ns ± 1%    +20.39%  (p=0.016 n=4+5)
DecompressDict/blockSize_100/level_5-8          24.7ns ± 0%    30.6ns ± 1%    +24.08%  (p=0.016 n=4+5)
DecompressDict/blockSize_100/level_10-8         23.5ns ± 0%    28.8ns ± 1%    +22.55%  (p=0.016 n=4+5)
DecompressDict/blockSize_1000/level_3-8          165ns ± 0%     175ns ± 0%     +6.00%  (p=0.016 n=4+5)
DecompressDict/blockSize_1000/level_5-8          416ns ± 1%     425ns ± 1%     +2.33%  (p=0.016 n=4+5)
DecompressDict/blockSize_1000/level_10-8         424ns ± 4%     421ns ± 1%       ~     (p=1.000 n=4+5)
DecompressDict/blockSize_10000/level_3-8        1.51µs ± 2%    1.51µs ± 1%       ~     (p=0.952 n=4+5)
DecompressDict/blockSize_10000/level_5-8        1.71µs ± 0%    1.72µs ± 1%     +0.62%  (p=0.032 n=4+5)
DecompressDict/blockSize_10000/level_10-8       1.61µs ± 0%    1.61µs ± 1%       ~     (p=0.952 n=4+5)
DecompressDict/blockSize_100000/level_3-8       13.8µs ± 1%    13.7µs ± 1%       ~     (p=0.286 n=4+5)
DecompressDict/blockSize_100000/level_5-8       13.5µs ± 1%    13.6µs ± 1%       ~     (p=1.000 n=4+5)
DecompressDict/blockSize_100000/level_10-8      12.5µs ± 1%    12.5µs ± 3%       ~     (p=0.730 n=4+5)
DecompressDict/blockSize_300000/level_3-8       37.8µs ± 1%    37.6µs ± 1%       ~     (p=0.413 n=4+5)
DecompressDict/blockSize_300000/level_5-8       44.2µs ± 1%    43.7µs ± 0%       ~     (p=0.095 n=4+5)
DecompressDict/blockSize_300000/level_10-8      36.4µs ± 1%    36.1µs ± 1%       ~     (p=0.190 n=4+5)
CompressDict/blockSize_1/level_3-8              38.9ns ± 0%    46.0ns ± 1%    +18.26%  (p=0.016 n=4+5)
CompressDict/blockSize_1/level_5-8              38.9ns ± 0%    45.9ns ± 3%    +17.95%  (p=0.016 n=4+5)
CompressDict/blockSize_1/level_10-8             39.5ns ± 1%    46.2ns ± 0%    +16.81%  (p=0.016 n=4+5)
CompressDict/blockSize_10/level_3-8              159ns ± 1%     167ns ± 2%     +5.36%  (p=0.016 n=4+5)
CompressDict/blockSize_10/level_5-8              170ns ± 1%     178ns ± 2%     +4.85%  (p=0.016 n=4+5)
CompressDict/blockSize_10/level_10-8             178ns ± 1%     185ns ± 1%     +4.29%  (p=0.016 n=4+5)
CompressDict/blockSize_100/level_3-8            81.0ns ± 0%    89.0ns ± 1%     +9.91%  (p=0.016 n=4+5)
CompressDict/blockSize_100/level_5-8             199ns ± 1%     206ns ± 1%     +3.51%  (p=0.016 n=4+5)
CompressDict/blockSize_100/level_10-8            186ns ± 1%     192ns ± 1%     +3.40%  (p=0.016 n=4+5)
CompressDict/blockSize_1000/level_3-8            399ns ± 1%     410ns ± 1%     +2.54%  (p=0.016 n=4+5)
CompressDict/blockSize_1000/level_5-8           1.26µs ± 1%    1.28µs ± 0%     +0.91%  (p=0.016 n=4+5)
CompressDict/blockSize_1000/level_10-8          2.33µs ± 2%    2.35µs ± 2%       ~     (p=0.905 n=4+5)
CompressDict/blockSize_10000/level_3-8          4.09µs ± 2%    4.10µs ± 1%       ~     (p=0.952 n=4+5)
CompressDict/blockSize_10000/level_5-8          10.1µs ± 0%    10.0µs ± 1%     -0.88%  (p=0.000 n=4+5)
CompressDict/blockSize_10000/level_10-8         24.4µs ± 1%    24.3µs ± 0%       ~     (p=0.190 n=4+5)
CompressDict/blockSize_100000/level_3-8         36.7µs ± 1%    36.6µs ± 1%       ~     (p=0.413 n=4+5)
CompressDict/blockSize_100000/level_5-8         94.3µs ± 1%    93.2µs ± 1%     -1.12%  (p=0.032 n=4+5)
CompressDict/blockSize_100000/level_10-8         159µs ± 1%     159µs ± 1%       ~     (p=0.730 n=4+5)
CompressDict/blockSize_300000/level_3-8          149µs ± 1%     149µs ± 1%       ~     (p=0.730 n=4+5)
CompressDict/blockSize_300000/level_5-8          346µs ± 1%     344µs ± 1%       ~     (p=0.413 n=4+5)
CompressDict/blockSize_300000/level_10-8        1.21ms ± 1%    1.19ms ± 1%     -1.45%  (p=0.016 n=4+5)
Compress/blockSize_1/level_3-8                  25.0ns ± 0%    29.8ns ± 0%    +19.10%  (p=0.016 n=4+5)
Compress/blockSize_1/level_5-8                  25.1ns ± 0%    29.8ns ± 0%    +18.95%  (p=0.016 n=4+5)
Compress/blockSize_1/level_10-8                 25.4ns ± 0%    30.3ns ± 1%    +19.24%  (p=0.016 n=4+5)
Compress/blockSize_10/level_3-8                 42.1ns ± 0%    48.1ns ± 0%    +14.33%  (p=0.016 n=4+5)
Compress/blockSize_10/level_5-8                 43.0ns ± 1%    48.8ns ± 1%    +13.27%  (p=0.016 n=4+5)
Compress/blockSize_10/level_10-8                44.7ns ± 2%    50.2ns ± 1%    +12.24%  (p=0.016 n=4+5)
Compress/blockSize_100/level_3-8                 172ns ± 1%     178ns ± 1%     +3.67%  (p=0.016 n=4+5)
Compress/blockSize_100/level_5-8                 236ns ± 0%     242ns ± 1%     +2.69%  (p=0.016 n=4+5)
Compress/blockSize_100/level_10-8                263ns ± 1%     270ns ± 1%     +2.81%  (p=0.016 n=4+5)
Compress/blockSize_1000/level_3-8                818ns ± 1%     821ns ± 1%       ~     (p=0.730 n=4+5)
Compress/blockSize_1000/level_5-8               1.22µs ± 1%    1.20µs ± 1%     -1.66%  (p=0.016 n=4+5)
Compress/blockSize_1000/level_10-8              1.72µs ± 1%    1.74µs ± 1%       ~     (p=0.079 n=4+5)
Compress/blockSize_10000/level_3-8              4.38µs ± 1%    4.41µs ± 1%       ~     (p=0.413 n=4+5)
Compress/blockSize_10000/level_5-8              9.54µs ± 1%    9.09µs ± 1%     -4.68%  (p=0.016 n=4+5)
Compress/blockSize_10000/level_10-8             39.6µs ± 3%    39.1µs ± 1%       ~     (p=0.730 n=4+5)
Compress/blockSize_100000/level_3-8             35.6µs ± 1%    35.2µs ± 3%       ~     (p=0.556 n=4+5)
Compress/blockSize_100000/level_5-8             89.5µs ± 1%    86.2µs ± 1%     -3.69%  (p=0.016 n=4+5)
Compress/blockSize_100000/level_10-8             462µs ± 0%     459µs ± 1%       ~     (p=0.190 n=4+5)
Compress/blockSize_300000/level_3-8              123µs ± 1%     122µs ± 1%     -1.10%  (p=0.032 n=4+5)
Compress/blockSize_300000/level_5-8              319µs ± 1%     322µs ± 3%       ~     (p=0.556 n=4+5)
Compress/blockSize_300000/level_10-8            1.08ms ± 1%    1.09ms ± 3%       ~     (p=0.730 n=4+5)
Decompress/blockSize_1/level_3-8                20.6ns ± 3%    23.8ns ± 3%    +15.18%  (p=0.016 n=4+5)
Decompress/blockSize_1/level_5-8                20.7ns ± 2%    23.7ns ± 4%    +14.34%  (p=0.016 n=4+5)
Decompress/blockSize_1/level_10-8               20.5ns ± 1%    23.2ns ± 4%    +13.09%  (p=0.016 n=4+5)
Decompress/blockSize_10/level_3-8               21.2ns ± 3%    23.4ns ± 3%    +10.56%  (p=0.016 n=4+5)
Decompress/blockSize_10/level_5-8               21.2ns ± 2%    23.8ns ± 4%    +11.96%  (p=0.016 n=4+5)
Decompress/blockSize_10/level_10-8              20.7ns ± 3%    23.6ns ± 2%    +14.14%  (p=0.016 n=4+5)
Decompress/blockSize_100/level_3-8              35.2ns ± 0%    39.9ns ± 0%    +13.37%  (p=0.016 n=4+5)
Decompress/blockSize_100/level_5-8              35.5ns ± 0%    40.0ns ± 1%    +12.54%  (p=0.016 n=4+5)
Decompress/blockSize_100/level_10-8             35.8ns ± 1%    40.2ns ± 1%    +12.27%  (p=0.016 n=4+5)
Decompress/blockSize_1000/level_3-8              449ns ± 1%     455ns ± 0%     +1.26%  (p=0.016 n=4+5)
Decompress/blockSize_1000/level_5-8              448ns ± 0%     458ns ± 1%     +2.03%  (p=0.016 n=4+5)
Decompress/blockSize_1000/level_10-8             444ns ± 0%     470ns ± 3%     +5.93%  (p=0.016 n=4+5)
Decompress/blockSize_10000/level_3-8            1.75µs ± 2%    1.74µs ± 1%       ~     (p=1.000 n=4+5)
Decompress/blockSize_10000/level_5-8            1.74µs ± 1%    1.74µs ± 0%       ~     (p=1.000 n=4+5)
Decompress/blockSize_10000/level_10-8           1.71µs ± 1%    1.73µs ± 1%       ~     (p=0.063 n=4+5)
Decompress/blockSize_100000/level_3-8           12.8µs ± 1%    12.5µs ± 0%     -2.27%  (p=0.029 n=4+4)
Decompress/blockSize_100000/level_5-8           14.7µs ± 1%    14.3µs ± 1%     -2.78%  (p=0.016 n=4+5)
Decompress/blockSize_100000/level_10-8          12.7µs ± 0%    12.5µs ± 0%     -2.06%  (p=0.016 n=4+5)
Decompress/blockSize_300000/level_3-8           39.0µs ± 1%    37.8µs ± 2%     -3.13%  (p=0.016 n=4+5)
Decompress/blockSize_300000/level_5-8           45.2µs ± 1%    43.8µs ± 0%     -3.23%  (p=0.016 n=4+5)
Decompress/blockSize_300000/level_10-8          37.0µs ± 1%    36.6µs ± 2%       ~     (p=0.111 n=4+5)
ReaderDict/blockSize_1/level_3-8                64.3ns ± 1%    58.1ns ± 3%     -9.62%  (p=0.016 n=4+5)
ReaderDict/blockSize_1/level_5-8                66.6ns ± 2%    57.6ns ± 1%    -13.54%  (p=0.016 n=4+5)
ReaderDict/blockSize_1/level_10-8               70.3ns ± 0%    63.9ns ± 1%     -9.08%  (p=0.016 n=4+5)
ReaderDict/blockSize_10/level_3-8               73.6ns ± 1%    68.7ns ± 2%     -6.66%  (p=0.016 n=4+5)
ReaderDict/blockSize_10/level_5-8               71.7ns ± 0%    66.3ns ± 1%     -7.54%  (p=0.016 n=4+5)
ReaderDict/blockSize_10/level_10-8              71.4ns ± 3%    64.8ns ± 1%     -9.22%  (p=0.016 n=4+5)
ReaderDict/blockSize_100/level_3-8               218ns ± 1%     211ns ± 1%     -3.15%  (p=0.016 n=4+5)
ReaderDict/blockSize_100/level_5-8               464ns ± 1%     462ns ± 1%       ~     (p=0.556 n=4+5)
ReaderDict/blockSize_100/level_10-8              459ns ± 1%     456ns ± 1%       ~     (p=0.190 n=4+5)
ReaderDict/blockSize_1000/level_3-8             1.62µs ± 0%    1.58µs ± 1%     -2.18%  (p=0.016 n=4+5)
ReaderDict/blockSize_1000/level_5-8             1.84µs ± 1%    1.78µs ± 1%     -3.29%  (p=0.016 n=4+5)
ReaderDict/blockSize_1000/level_10-8            1.73µs ± 1%    1.71µs ± 8%       ~     (p=0.190 n=4+5)
ReaderDict/blockSize_10000/level_3-8            14.9µs ± 1%    14.5µs ± 2%     -2.63%  (p=0.016 n=4+5)
ReaderDict/blockSize_10000/level_5-8            14.8µs ± 1%    14.4µs ± 2%     -2.36%  (p=0.016 n=4+5)
ReaderDict/blockSize_10000/level_10-8           13.7µs ± 0%    13.2µs ± 2%     -3.62%  (p=0.016 n=4+5)
ReaderDict/blockSize_100000/level_3-8            149µs ± 2%     138µs ± 1%     -6.94%  (p=0.016 n=4+5)
ReaderDict/blockSize_100000/level_5-8            159µs ± 2%     150µs ± 1%     -5.74%  (p=0.016 n=4+5)
ReaderDict/blockSize_100000/level_10-8           131µs ± 1%     123µs ± 1%     -5.78%  (p=0.016 n=4+5)
ReaderDict/blockSize_300000/level_3-8            511µs ± 1%     479µs ± 2%     -6.29%  (p=0.016 n=4+5)
ReaderDict/blockSize_300000/level_5-8            531µs ± 1%     494µs ± 1%     -6.93%  (p=0.016 n=4+5)
ReaderDict/blockSize_300000/level_10-8           415µs ± 1%     396µs ± 1%     -4.57%  (p=0.016 n=4+5)
Reader/blockSize_1/level_3-8                    63.6ns ± 1%    56.6ns ± 2%    -10.96%  (p=0.016 n=4+5)
Reader/blockSize_1/level_5-8                    64.2ns ± 1%    57.1ns ± 1%    -11.03%  (p=0.016 n=4+5)
Reader/blockSize_1/level_10-8                   66.1ns ± 4%    57.4ns ± 1%    -13.14%  (p=0.016 n=4+5)
Reader/blockSize_10/level_3-8                   83.4ns ± 9%    75.1ns ± 1%     -9.92%  (p=0.016 n=4+5)
Reader/blockSize_10/level_5-8                   81.4ns ± 1%    75.9ns ± 3%     -6.80%  (p=0.016 n=4+5)
Reader/blockSize_10/level_10-8                  83.3ns ± 5%    76.2ns ± 1%     -8.57%  (p=0.016 n=4+5)
Reader/blockSize_100/level_3-8                   496ns ± 1%     491ns ± 1%       ~     (p=0.190 n=4+5)
Reader/blockSize_100/level_5-8                   496ns ± 1%     491ns ± 1%       ~     (p=0.111 n=4+5)
Reader/blockSize_100/level_10-8                  490ns ± 1%     487ns ± 0%     -0.71%  (p=0.016 n=4+5)
Reader/blockSize_1000/level_3-8                 1.87µs ± 0%    1.82µs ± 1%     -2.47%  (p=0.016 n=4+5)
Reader/blockSize_1000/level_5-8                 1.87µs ± 0%    1.83µs ± 0%     -2.19%  (p=0.016 n=4+5)
Reader/blockSize_1000/level_10-8                1.85µs ± 0%    1.80µs ± 1%     -2.61%  (p=0.016 n=4+5)
Reader/blockSize_10000/level_3-8                13.9µs ± 1%    13.6µs ± 2%       ~     (p=0.190 n=4+5)
Reader/blockSize_10000/level_5-8                15.9µs ± 2%    15.5µs ± 2%       ~     (p=0.190 n=4+5)
Reader/blockSize_10000/level_10-8               13.8µs ± 1%    13.5µs ± 2%       ~     (p=0.111 n=4+5)
Reader/blockSize_100000/level_3-8                149µs ± 1%     141µs ± 3%     -5.23%  (p=0.016 n=4+5)
Reader/blockSize_100000/level_5-8                162µs ± 1%     153µs ± 2%     -5.70%  (p=0.016 n=4+5)
Reader/blockSize_100000/level_10-8               132µs ± 2%     125µs ± 1%     -5.22%  (p=0.016 n=4+5)
Reader/blockSize_300000/level_3-8                514µs ± 1%     480µs ± 1%     -6.62%  (p=0.016 n=4+5)
Reader/blockSize_300000/level_5-8                536µs ± 2%     502µs ± 1%     -6.26%  (p=0.016 n=4+5)
Reader/blockSize_300000/level_10-8               422µs ± 1%     403µs ± 4%       ~     (p=0.063 n=4+5)
StreamCompress/blockSize_1/level_3-8            6.36µs ± 2%    6.32µs ± 2%       ~     (p=0.365 n=4+5)
StreamCompress/blockSize_1/level_5-8            47.7µs ± 0%    48.1µs ± 1%     +0.87%  (p=0.032 n=4+5)
StreamCompress/blockSize_1/level_10-8            378µs ± 1%     380µs ± 1%       ~     (p=0.190 n=4+5)
StreamCompress/blockSize_10/level_3-8           6.19µs ± 0%    6.22µs ± 0%       ~     (p=0.063 n=4+5)
StreamCompress/blockSize_10/level_5-8           47.6µs ± 1%    48.4µs ± 1%     +1.67%  (p=0.016 n=4+5)
StreamCompress/blockSize_10/level_10-8           377µs ± 1%     379µs ± 0%       ~     (p=0.111 n=4+5)
StreamCompress/blockSize_100/level_3-8          5.66µs ± 2%    5.66µs ± 1%       ~     (p=1.000 n=4+5)
StreamCompress/blockSize_100/level_5-8          46.5µs ± 1%    46.3µs ± 1%       ~     (p=0.905 n=4+5)
StreamCompress/blockSize_100/level_10-8          379µs ± 1%     379µs ± 0%       ~     (p=1.000 n=4+5)
StreamCompress/blockSize_1000/level_3-8         5.39µs ± 2%    5.42µs ± 1%       ~     (p=0.413 n=4+5)
StreamCompress/blockSize_1000/level_5-8         36.1µs ± 3%    36.7µs ± 4%       ~     (p=0.556 n=4+5)
StreamCompress/blockSize_1000/level_10-8         389µs ± 1%     406µs ± 3%     +4.28%  (p=0.016 n=4+5)
StreamCompress/blockSize_10000/level_3-8        38.9µs ± 2%    37.3µs ± 1%     -4.09%  (p=0.016 n=4+5)
StreamCompress/blockSize_10000/level_5-8         112µs ± 2%     111µs ± 3%       ~     (p=0.413 n=4+5)
StreamCompress/blockSize_10000/level_10-8        707µs ± 1%     711µs ± 2%       ~     (p=0.413 n=4+5)
StreamCompress/blockSize_100000/level_3-8        535µs ± 1%     512µs ± 2%     -4.34%  (p=0.016 n=4+5)
StreamCompress/blockSize_100000/level_5-8       1.46ms ± 2%    1.41ms ± 2%     -3.44%  (p=0.016 n=4+5)
StreamCompress/blockSize_100000/level_10-8      5.04ms ± 2%    5.02ms ± 1%       ~     (p=0.905 n=4+5)
StreamCompress/blockSize_300000/level_3-8       1.88ms ± 1%    1.82ms ± 1%     -3.13%  (p=0.016 n=4+5)
StreamCompress/blockSize_300000/level_5-8       5.63ms ± 3%    5.54ms ± 4%       ~     (p=0.111 n=4+5)
StreamCompress/blockSize_300000/level_10-8      18.4ms ± 9%    17.8ms ± 1%       ~     (p=0.730 n=4+5)
StreamDecompress/blockSize_1/level_3-8          68.6ns ± 3%    57.4ns ± 1%    -16.40%  (p=0.016 n=4+5)
StreamDecompress/blockSize_1/level_5-8          70.9ns ± 5%    60.6ns ± 9%    -14.47%  (p=0.016 n=4+5)
StreamDecompress/blockSize_1/level_10-8         71.8ns ± 6%    61.3ns ±13%    -14.62%  (p=0.016 n=4+5)
StreamDecompress/blockSize_10/level_3-8         87.1ns ± 6%    81.7ns ±11%       ~     (p=0.286 n=4+5)
StreamDecompress/blockSize_10/level_5-8         88.9ns ± 6%    76.9ns ± 1%    -13.55%  (p=0.029 n=4+4)
StreamDecompress/blockSize_10/level_10-8        88.0ns ± 4%    80.0ns ± 6%       ~     (p=0.063 n=4+5)
StreamDecompress/blockSize_100/level_3-8         500ns ± 1%     489ns ± 0%     -2.22%  (p=0.016 n=4+5)
StreamDecompress/blockSize_100/level_5-8         493ns ± 0%     491ns ± 1%       ~     (p=0.302 n=4+5)
StreamDecompress/blockSize_100/level_10-8        487ns ± 1%     487ns ± 1%       ~     (p=1.000 n=4+5)
StreamDecompress/blockSize_1000/level_3-8       1.86µs ± 1%    1.82µs ± 0%     -2.14%  (p=0.029 n=4+4)
StreamDecompress/blockSize_1000/level_5-8       1.86µs ± 0%    1.84µs ± 1%     -1.17%  (p=0.032 n=4+5)
StreamDecompress/blockSize_1000/level_10-8      1.84µs ± 1%    1.83µs ± 1%       ~     (p=0.286 n=4+5)
StreamDecompress/blockSize_10000/level_3-8      13.6µs ± 2%    13.7µs ± 3%       ~     (p=0.794 n=4+5)
StreamDecompress/blockSize_10000/level_5-8      15.6µs ± 1%    15.1µs ± 1%     -2.70%  (p=0.016 n=4+5)
StreamDecompress/blockSize_10000/level_10-8     13.4µs ± 0%    13.2µs ± 1%     -1.41%  (p=0.016 n=4+5)
StreamDecompress/blockSize_100000/level_3-8      144µs ± 1%     141µs ± 1%     -2.29%  (p=0.016 n=4+5)
StreamDecompress/blockSize_100000/level_5-8      157µs ± 1%     153µs ± 2%       ~     (p=0.063 n=4+5)
StreamDecompress/blockSize_100000/level_10-8     128µs ± 1%     126µs ± 1%     -1.72%  (p=0.016 n=4+5)
StreamDecompress/blockSize_300000/level_3-8      492µs ± 0%     482µs ± 1%     -1.98%  (p=0.016 n=4+5)
StreamDecompress/blockSize_300000/level_5-8      510µs ± 0%     503µs ± 1%     -1.37%  (p=0.016 n=4+5)
StreamDecompress/blockSize_300000/level_10-8     404µs ± 0%     398µs ± 1%     -1.45%  (p=0.016 n=4+5)
WriterDict/blockSize_1/level_3-8                 255ns ± 1%     264ns ± 1%     +3.72%  (p=0.016 n=4+5)
WriterDict/blockSize_1/level_5-8                 297ns ± 1%     309ns ± 2%     +4.03%  (p=0.016 n=4+5)
WriterDict/blockSize_1/level_10-8                302ns ± 1%     315ns ± 1%     +4.27%  (p=0.016 n=4+5)
WriterDict/blockSize_10/level_3-8                180ns ± 1%     191ns ± 0%     +5.99%  (p=0.016 n=4+5)
WriterDict/blockSize_10/level_5-8                326ns ± 1%     335ns ± 1%     +2.66%  (p=0.016 n=4+5)
WriterDict/blockSize_10/level_10-8               312ns ± 2%     317ns ± 1%       ~     (p=0.190 n=4+5)
WriterDict/blockSize_100/level_3-8               502ns ± 2%     511ns ± 1%       ~     (p=0.111 n=4+5)
WriterDict/blockSize_100/level_5-8              1.46µs ± 1%    1.44µs ± 1%     -1.16%  (p=0.032 n=4+5)
WriterDict/blockSize_100/level_10-8             2.59µs ± 1%    2.53µs ± 1%     -2.28%  (p=0.016 n=4+5)
WriterDict/blockSize_1000/level_3-8             4.59µs ± 1%    4.45µs ± 2%     -3.02%  (p=0.016 n=4+5)
WriterDict/blockSize_1000/level_5-8             11.6µs ± 2%    11.0µs ± 1%     -4.56%  (p=0.016 n=4+5)
WriterDict/blockSize_1000/level_10-8            27.3µs ± 1%    26.1µs ± 1%     -4.55%  (p=0.016 n=4+5)
WriterDict/blockSize_10000/level_3-8            39.8µs ± 0%    38.1µs ± 1%     -4.37%  (p=0.016 n=4+5)
WriterDict/blockSize_10000/level_5-8             116µs ± 2%     110µs ± 1%     -4.87%  (p=0.016 n=4+5)
WriterDict/blockSize_10000/level_10-8            244µs ± 2%     232µs ± 2%     -4.72%  (p=0.016 n=4+5)
WriterDict/blockSize_100000/level_3-8            409µs ± 2%     393µs ± 0%     -3.70%  (p=0.016 n=4+5)
WriterDict/blockSize_100000/level_5-8           1.06ms ± 2%    1.02ms ± 1%     -3.63%  (p=0.016 n=4+5)
WriterDict/blockSize_100000/level_10-8          2.26ms ± 3%    2.17ms ± 2%     -3.97%  (p=0.032 n=4+5)
…